--- packages/router/src/router/redirect.ts
@@ -1,7 +1,0 @@
-import { ExternalRedirect, RedirectLocation, RedirectProps } from "@curi/types";
-
-export function isExternalRedirect(
-  redirect: ExternalRedirect | RedirectLocation | RedirectProps
-): redirect is ExternalRedirect {
-  return "externalURL" in redirect;
-}
