--- packages/router/types/router/redirect.d.ts
@@ -1,2 +1,0 @@
-import { ExternalRedirect, RedirectLocation, RedirectProps } from "@curi/types";
-export declare function isExternalRedirect(redirect: ExternalRedirect | RedirectLocation | RedirectProps): redirect is ExternalRedirect;
