+++ packages/router/types/router/typeGuards.d.ts
@@ -1,0 +1,3 @@
+import { Route, AsyncRoute, ExternalRedirect, RedirectLocation, RedirectProps } from "@curi/types";
+export declare function isAsyncRoute(route: Route): route is AsyncRoute;
+export declare function isExternalRedirect(redirect: ExternalRedirect | RedirectLocation | RedirectProps): redirect is ExternalRedirect;
